Create  PrimaryItemForNotification
