+++ apps/app/src/components/Sidebar/SidebarNav/SidebarNav.tsx
@@ -1,7 +1,5 @@
 import React, { memo } from 'react';
 
-import dynamic from 'next/dynamic';
-
 import { SidebarContentsType } from '~/interfaces/ui';
 
 import { PageCreateButton } from '../PageCreateButton';
@@ -10,8 +8,6 @@ import { PrimaryItems } from './PrimaryItems';
 import { SecondaryItems } from './SecondaryItems';
 
 import styles from './SidebarNav.module.scss';
-
-const PrimaryItemsForNotification = dynamic(() => import('./PrimaryItems').then(mod => mod.PrimaryItemsForNotification), { ssr: false });
 
 export type SidebarNavProps = {
   onPrimaryItemHover?: (contents: SidebarContentsType) => void,
@@ -26,7 +22,6 @@ export const SidebarNav = memo((props: SidebarNavProps) => {
 
       <div className="grw-sidebar-nav-primary-container" data-vrt-blackout-sidebar-nav>
         <PrimaryItems onItemHover={onPrimaryItemHover} />
-        <PrimaryItemsForNotification onItemHover={onPrimaryItemHover} />
       </div>
       <div className="grw-sidebar-nav-secondary-container">
         <SecondaryItems />
